Support renewal for same level, different gateway.
* ENHANCEMENT: Support members renewing while their check order is pending. Helps support gateway switches when navigating to checkout directly and changing payment methods.

This will cancel all previous orders that are pending and linked to the 'check' gateway. Do not cancel if the user is purchasing a different level (MMPU) or if the level settings are not configured for PBC. (_MIGHT_ be an issue if you toggle this on/off and users signup and change).